Alter:DB structure for efficiency

+++ src/main/java/com/docnow/docnow/auth/repository/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<UserData,Integer> {
+public interface UserRepository extends JpaRepository<UserData,Long> {
     Boolean existsByUsername(String username) ;
     Optional<UserData> findByUsername(String username);
 }
